WF/requirements.txt
pandas==2.2.2
openpyxl==3.1.2
pymysql==1.1.0
sqlalchemy==2.0.29
python-dotenv==1.0.1